requirements.txt
django-ninja==1.1.0
uvicorn==0.29.0
psycopg2-binary
openai==1.16.2